cleanup

+++ test/flow/script/run-snap-tests.sh
@@ -4,19 +4,18 @@
 # set -e
 # set -x
 
-
 # Kills process tree
 killtree() {
-  local _pid=$1
-  kill -stop ${_pid} # needed to stop quickly forking parent from producing children between child killing and parent killing
-  for _child in $(pgrep -P ${_pid}); do
-    killtree ${_child}
-  done
-  kill -TERM ${_pid}
+  if [ -n "$1" ]; then
+    local _pid=$1
+    kill -stop ${_pid} # needed to stop quickly forking parent from producing children between child killing and parent killing
+    for _child in $(pgrep -P ${_pid}); do
+      killtree ${_child}
+    done
+    kill -TERM ${_pid}
+  fi
 }
 
-
-clear
 echo "Running ui-tests"
 echo "***************************"
 
@@ -39,19 +38,12 @@ DBUS_SESSION_BUS_ADDRESS=/dev/null CHROMEDRIVER_VERSION=2.24 $CHROMEDRIVER --ver
 DRIVER_PID=$!
 sleep 4
 
-#while true; do ps aux;sleep 10;done &
-#WATCHER_PID=$!
-
 echo "Running tests"
 $NIGHTWATCH_BINARY -c ./test/flow/nightwatch-snap.json --retries 3
 TESTSTATUS=$?
 echo "Done"
 
-
-#killtree $WATCHER_PID
-if [ -n "$NODE_PID" ]; then
-  killtree $NODE_PID
-fi
+killtree $NODE_PID
 killtree $DRIVER_PID
 
 echo "Exiting with status $TESTSTATUS"
